feat(drainer): polished the code

+++ crates/router/src/configs/settings.rs
@@ -144,7 +144,7 @@ pub struct DrainerSettings {
     pub stream_name: String,
     pub num_partitions: u8,
     pub batch_size: usize,
-    pub max_read_count: usize,
+    pub max_read_count: u64,
 }
 
 impl Settings {
